+++ BuildMasterAutomation/Functions/Invoke-BMVariableEndpoint.ps1
@@ -12,10 +12,11 @@ function Invoke-BMVariableEndpoint
         [Object] $Variable,
 
         [Parameter(Mandatory, ParameterSetName='Set')]
+        [AllowEmptyString()]
         [String] $Value,
 
         [Parameter(Mandatory)]
-        [ValidateSet('application', 'application-group', 'environment', 'global', 'server', 'role')]
+        [ValidateSet('application', 'application-group', 'environment', 'global', 'server', 'role', 'releases', 'builds')]
         [String] $EntityTypeName,
 
         [Parameter(Mandatory)]
@@ -58,21 +59,25 @@ function Invoke-BMVariableEndpoint
     if ($EntityTypeName -ne 'global')
     {
         $entityTypeDescriptions = @{
-            'application' = 'application';
+            'application'       = 'application';
             'application-group' = 'application group';
-            'environment' = 'environment';
-            'server' = 'server';
-            'role' = 'server role'
+            'environment'       = 'environment';
+            'server'            = 'server';
+            'role'              = 'server role';
+            'releases'          = 'release';
+            'builds'            = 'build';
         }
         $entityDesc = $entityTypeDescriptions[$EntityTypeName]
         $entityDescCapitalized = [char]::ToUpperInvariant($entityDesc[0]) + $entityDesc.Substring(1)
 
         $entityToParamNameMap = @{
-            'application' = 'Application';
+            'application'       = 'Application';
             'application-group' = 'ApplicationGroup';
-            'environment' = 'Environment';
-            'server' = 'Server';
-            'role' = 'ServerRole';
+            'environment'       = 'Environment';
+            'server'            = 'Server';
+            'role'              = 'ServerRole';
+            'releases'          = 'Release';
+            'builds'            = 'Build';
         }
 
         # What parameter has the variable's entity?
@@ -84,31 +89,55 @@ function Invoke-BMVariableEndpoint
         $getEntityArg = @{
             $paramName = $entity;
         }
-        # Check if the entity exists in BuildMaster.
-        $bmEntity = & "Get-BM$($paramName)" -Session $Session @getEntityArg -ErrorAction Ignore
-        if (-not $bmEntity)
+
+        if ($EntityTypeName -in @('releases', 'builds'))
         {
-            $entityName = $entity | Get-BMObjectName
-            $msg = "$($entityDescCapitalized) ""$($entityName)"" does not exist."
-            if ($deleting)
+            if ($entity.GetType().Name -ne 'PSCustomObject')
             {
-                $msg = "Unable to delete variable ""$($variableName)"" because the $($entityDesc) " +
-                       """$($entityName)"" does not exist."
+                $msg = "The ${paramName} parameter must be a ${paramName} object from the Get-BM${paramName} " +
+                       "function, but it was a $($entity.GetType().Name) ""${entity}""."
+                Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+                return
             }
-            elseif ($updating)
+
+            # It's required to pass in an object for variables on Releases and Builds, therefore we know those already
+            # exists and can use the objects directly.
+            $entityPathSegment = "${EntityTypeName}/$([Uri]::EscapeDataString($entity.applicationName))/$([Uri]::EscapeDataString($entity.releaseNumber))"
+
+            if ($EntityTypeName -eq 'builds')
             {
-                $msg = "Unable to set variable ""$($variableName)"" because the $($entityDesc) ""$($entityName)"" " +
-                       'does not exist.'
+                $entityPathSegment = "${entityPathSegment}/$([Uri]::EscapeDataString($entity.buildNumber))"
             }
-            Write-Error -Message $msg -ErrorAction $ErrorActionPreference
-            return
+
+            $entityPathSegment = "${entityPathSegment}${variablePathSegment}"
         }
+        else
+        {
+            # Check if the entity exists in BuildMaster.
+            $bmEntity = & "Get-BM$($paramName)" -Session $Session @getEntityArg -ErrorAction Ignore
+            if (-not $bmEntity)
+            {
+                $entityName = $entity | Get-BMObjectName
+                $msg = "$($entityDescCapitalized) ""$($entityName)"" does not exist."
+                if ($deleting)
+                {
+                    $msg = "Unable to delete variable ""$($variableName)"" because $($entityDesc) " +
+                        """$($entityName)"" does not exist."
+                }
+                elseif ($updating)
+                {
+                    $msg = "Unable to set variable ""$($variableName)"" because $($entityDesc) ""$($entityName)"" " +
+                        'does not exist.'
+                }
+                Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+                return
+            }
 
-        # Get the entity's name.
-        $entityName = $bmEntity | Get-BMOBjectName -Strict -ObjectTypeName $paramName
+            $entityName = $bmEntity | Get-BMOBjectName -Strict -ObjectTypeName $paramName
 
-        # Create the entity-specific endpoint path.
-        $entityPathSegment = "$($entityTypeName)/$([Uri]::EscapeDataString($entityName))$($variablePathSegment)"
+            # Create the entity-specific endpoint path.
+            $entityPathSegment = "$($EntityTypeName)/$([Uri]::EscapeDataString($entityName))$($variablePathSegment)"
+        }
     }
 
     $endpointPath = "variables/$($entityPathSegment)"
@@ -127,8 +156,10 @@ function Invoke-BMVariableEndpoint
         $nativeApiEntityIdParam['ApplicationGroup_Id'] = $bmEntity.ApplicationGroup_Id
     }
 
-    if (-not $updating)
+    if ($getting -or $deleting)
     {
+        $variableExists = $false
+
         if ($useNativeApi)
         {
             $nativeVariables = Invoke-BMNativeApiMethod -Session $Session `
@@ -138,6 +169,7 @@ function Invoke-BMVariableEndpoint
 
             foreach ($nativeVar in $nativeVariables)
             {
+                $variableExists = $true
                 $bytes = [Convert]::FromBase64String($nativeVar.Variable_Value)
                 $variables[$nativeVar.Variable_Name] = [Text.Encoding]::UTF8.GetString($bytes)
             }
@@ -145,30 +177,39 @@ function Invoke-BMVariableEndpoint
         }
         else
         {
-            $variables = Invoke-BMRestMethod -Session $session -Name $endpointPath
-        }
-    }
-
-    if ($Variable -and -not $searching -and -not $variables -and -not $updating)
-    {
-        $msg = "Variable ""$($variableName)"" does not exist."
-        if ($bmEntity)
-        {
-            $msg = "$($entityDescCapitalized) ""$($entityName)"" variable ""$($variableName)"" does not exist."
-        }
-
-        if ($ForDelete)
-        {
-            $msg = "Unable to delete variable ""$($variableName)"" because it does not exist."
-            if ($bmEntity)
+            try
             {
-                $msg = "Unable to delete $($entityDesc) ""$($entityName)"" variable ""$($variableName)"" because the " +
-                       "variable does not exist."
+                $variables = Invoke-BMRestMethod -Session $session -Name $endpointPath -ErrorAction Stop
+                $variableExists = $true
+            }
+            catch
+            {
+                $Global:Error.RemoveAt(0)
+                Write-Error -ErrorRecord $_ -ErrorAction $ErrorActionPreference
             }
         }
 
-        Write-Error -Message $msg -ErrorAction $ErrorActionPreference
-        return
+        if (-not $searching -and -not $variableExists)
+        {
+            $msg = "Variable ""$($variableName)"" does not exist."
+            if ($bmEntity)
+            {
+                $msg = "$($entityDescCapitalized) ""$($entityName)"" variable ""$($variableName)"" does not exist."
+            }
+
+            if ($deleting)
+            {
+                $msg = "Unable to delete variable ""$($variableName)"" because it does not exist."
+                if ($bmEntity)
+                {
+                    $msg = "Unable to delete $($entityDesc) ""$($entityName)"" variable ""$($variableName)"" because the " +
+                        "variable does not exist."
+                }
+            }
+
+            Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+            return
+        }
     }
 
     if ($deleting)
@@ -191,7 +232,12 @@ function Invoke-BMVariableEndpoint
         return
     }
 
-    if ($variables -is [String])
+    if ($variables -eq $null)
+    {
+        $variables = ''
+    }
+
+    if ($variables.GetType().Name -ne 'PSCustomObject')
     {
         return [pscustomobject]@{
             Name = $variableName;